Moved environments to their own file, added Production and Development accounts
