--- src/DeploymentEnvironment.ts
@@ -1,7 +1,0 @@
-import { Environment } from 'aws-cdk-lib';
-
-export interface DeploymentEnvironment {
-  accountName: string;
-  env: Environment;
-  assumedRolesToAlarmOn?: string|string[];
-}
