+++ src/main.ts
@@ -1,9 +1,9 @@
 import { App } from 'aws-cdk-lib';
-import { DeploymentEnvironment } from './DeploymentEnvironment';
+import { deploymentEnvironments } from './DeploymentEnvironments';
 import { PipelineStack } from './PipelineStack';
 
 // for development, use sandbox account
-const deploymentEnvironment = {
+const deployFromEnvironment = {
   account: '418648875085',
   region: 'eu-west-1',
 };
@@ -13,52 +13,12 @@ const sandboxEnvironment = {
   region: 'eu-west-1',
 };
 
-// const acceptanceEnvironment = {
-// 	account: '315037222840',
-// 	region: 'eu-west-1',
-// };
-
-// const productionEnvironment = {
-// 	account: '196212984627',
-// 	region: 'eu-west-1',
-// };
-
 const app = new App();
-
-
-/**
- * List all environments for which which a monitoring
- * pipeline should be deployed in prod
- */
-const deploymentEnvironments: DeploymentEnvironment[] = [
-  {
-    accountName: 'auth-accp',
-    env: {
-      account: '315037222840',
-      region: 'eu-west-1',
-    },
-  },
-  {
-    accountName: 'auth-prod',
-    env: {
-      account: '196212984627',
-      region: 'eu-west-1',
-    },
-  },
-  {
-    accountName: 'dns',
-    env: {
-      account: '108197740505',
-      region: 'eu-west-1',
-    },
-    assumedRolesToAlarmOn: 'nijmegen-operator',
-  },
-];
 
 if ('BRANCH_NAME' in process.env == false || process.env.BRANCH_NAME == 'development') {
   new PipelineStack(app, 'aws-monitoring-pipeline-development',
     {
-      env: deploymentEnvironment,
+      env: deployFromEnvironment,
       branchName: 'development',
       deployToEnvironments: [{ accountName: 'sandbox', env: sandboxEnvironment, assumedRolesToAlarmOn: 'Developers' }],
       environmentName: 'sandbox',
@@ -67,7 +27,7 @@ if ('BRANCH_NAME' in process.env == false || process.env.BRANCH_NAME == 'develop
 } else if ( process.env.BRANCH_NAME == 'main') {
   new PipelineStack(app, 'aws-monitoring-pipeline-prod',
     {
-      env: deploymentEnvironment,
+      env: deployFromEnvironment,
       branchName: 'main',
       deployToEnvironments: deploymentEnvironments,
       environmentName: 'production',
